requirements-dev.txt
asgiref==3.4.1
astroid==2.5.1
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.5
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.5.2
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
pylint==2.7.2
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
wincertstore==0.2
wrapt==1.12.1
django-allauth==0.46.0